Fixed an issue with read <poem name> and updated help
For read <poem name>, we use the search function of the Poetry Foundation website. Previously, we did not restrict the types of pages it searched for, which allowed it to select author pages as if they were poems. This is no longer the case.

Poetry Reader was put into the help as PoetryReader or poetryreader a couple times. These have been updated.

Version now v1.2.1